GMT MIRROR GENERICS
Represent M1 and M2 GMT mirrors with the struct
`Mirror<M>` where `M` is either `gmt_m1` or `gmt_m2`
